+++ src/iceqube/scheduler/classes.py
@@ -4,7 +4,7 @@ from iceqube.common.six.moves.queue import Full
 
 from iceqube.common.utils import InfiniteLoopThread
 from iceqube.messaging.classes import Message, MessageType, CancelMessage
-from iceqube.threading import Event
+from iceqube.async import Event
 
 
 class Scheduler(object):
