+++ src/iceqube/storage/backends/default.py
@@ -2,7 +2,7 @@ import abc
 from collections import defaultdict
 
 from iceqube.exceptions import TimeoutError
-from iceqube.threading import Event
+from iceqube.async import Event
 
 JOB_EVENT_MAPPING = defaultdict(lambda: Event())
