+++ tests/common/test_classes.py
@@ -8,7 +8,7 @@ from iceqube.client import SimpleClient
 from iceqube.common.classes import Job, State
 from iceqube.common.utils import import_stringified_func, stringify_func
 from iceqube.storage.backends import inmem
-from iceqube.threading import Event
+from iceqube.async import Event
 
 
 @pytest.fixture
