+++ tests/worker/backends/test_inmem.py
@@ -3,7 +3,7 @@ import pytest
 from iceqube.common.classes import Job
 from iceqube.messaging.backends.inmem import MessagingBackend
 from iceqube.messaging.classes import MessageType
-from iceqube.threading import Event
+from iceqube.async import Event
 from iceqube.worker.backends import inmem
